+++ app/components/comparison/shared/components/table-data.set.ts
@@ -55,11 +55,20 @@ export class TableDataSet {
                     }
                 }
             }
+            const foregroundColors: ColorDictionary = new ColorDictionary();
+            if (obj.type && obj.type.values) {
+                for (const v of obj.type.values) {
+                    if (v.foreground) {
+                        foregroundColors.setColor(v.name, v.foreground);
+                    }
+                }
+            }
             const type: Type = new Type(
                 obj.type.tag,
                 obj.type.class,
                 lcls,
-                colors
+                colors,
+                foregroundColors
             );
             let order = obj.order;
             if (!isNullOrUndefined(order)) {