+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
 - Enabling and disabling elements
 - Clicking labels for filtering
@@ -13,7 +14,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - replaced angular2-select with ng2-select
 
-## [1.0.0] - unreleased
+### Fixed
+- Improved foreground colors when auto assigning
+- Fixed problem with sorting `repo` columns [#114](https://github.com/ultimate-comparisons/ultimate-comparison-BASE/issues/114)
+- Add expand and shrink option for table [#90](https://github.com/ultimate-comparisons/ultimate-comparison-BASE/issues/90)
+
+## 1.0.0 - 2017-08-24
 ### Added
 - This file
 - Static VersionInformation class
@@ -21,6 +27,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added force push for deployment
 - Moved deployment to .travis.yml
 - Added tslint
+- Added unknown label for empty values
+- Enabling and disabling elements
 
 ### Fixed
 - Rate Limit exceeded for GitHub for up to 60 elements.